matchbox/matchbox-window-manager: Remove unnecessary dependencies on libmatchbox and libmatchbox-dev in depends and makedepends respectively, and enable PNG themes.
Updated programs and utilities related to Matchbox no longer make use of
libmatchbox as a shared library. This applies equally to Matchbox Window
Manager. libmatchbox is neither a build nor runtime dependency, so remove it.

A small patch is enclosed to enable PNG theme support to build correctly. The
'--enable-png-theme' option was added to the arguments of 'configure' to build
this.