matchbox/matchbox-common: Remove unnecessary dependency on libmatchbox-dev
matchbox-common previously checked for PNG support inlibmatchbox.pc.
This is unnecessary because it is now the only package that uses libmatchbox.

The solution is to add an option enable-png-icons toits configure script, so
that we can enable the building of PNG icons (in /usr/share/pixmaps) at
build-time,rather than have the configure script check.